Successful assembly
Added recursive matching function to search with matchPairs(), added
public Match() function to hide the implementation.

This gives a successful assembly of the 4-sequence example, and more
importantly assembles the 50-sequence sample data.  Verified the
50-sequence data is correct using:
http://www.genome.jp/tools/egassembler/
to assemble the FASTA sample data.  Currently the output is unbroken,
but `fold -w60` will match the egassembler data, and shows an identical
assembly.

There's still some debug info in the recursion, plus some cleanup and
tests to write, but now is a good time to snapshot the code.