Custom error type in assemble, refactor recurseMatch
recurseMatch was using a bool to determine completion, and swallowing errors.
That was not very idiomatic, and hid useful info. Refactored to use the error to
determine success, cleaned up some recursion logic.

Added a "placeError" type to assemble() and a check in recurseMatch() to
log simple placement errors if debug is enabled (normal behavior in searching
the list), and abort on any more serious errors.